refactor: tidy up discussion class
